+++ apps/kokomen-webview/src/domains/dashboard/components/interviewHistory.tsx
@@ -1,7 +1,7 @@
 import { getInterviewHistory } from "@/domains/dashboard/api";
 import { useIntersectionObserver } from "@kokomen/utils/react/useIntersectionObserver";
 import { interviewHistoryKeys } from "@kokomen/utils/general/querykeys";
-import Select from "@kokomen/ui/components/select";
+import { Select } from "@kokomen/ui";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import {
   Calendar,
