+++ apps/kokomen-webview/src/domains/members/components/memberQuestionFeedback.tsx
@@ -1,9 +1,9 @@
 import { useMutation } from "@tanstack/react-query";
 import { Heart } from "lucide-react";
 import { useMemo, useState } from "react";
-import { useToast } from "@kokomen/ui/hooks/useToast";
+import { useToast } from "@kokomen/ui";
 import { isAxiosError } from "axios";
-import { Button } from "@kokomen/ui/components/button";
+import { Button } from "@kokomen/ui";
 import { toggleMemberInterviewAnswerLike } from "@/domains/members/api";
 import { CamelCasedProperties } from "@kokomen/utils/general/convertConvention";
 import { MemberInterviewResult } from "@kokomen/types";
