+++ apps/kokomen-webview/src/domains/members/components/memberTotalFeedback.tsx
@@ -2,9 +2,9 @@ import { toggleMemberInterviewLike } from "@/domains/members/api";
 import { useMutation } from "@tanstack/react-query";
 import { CheckCircle, Eye, Heart, MessageCircle, Trophy } from "lucide-react";
 import { JSX, useState } from "react";
-import { useToast } from "@kokomen/ui/hooks/useToast";
+import { useToast } from "@kokomen/ui";
 import { isAxiosError } from "axios";
-import { Button } from "@kokomen/ui/components/button";
+import { Button } from "@kokomen/ui";
 import { MemberInterviewResult } from "@kokomen/types";
 import { CamelCasedProperties } from "@kokomen/utils/general/convertConvention";
 // import { captureButtonEvent } from "@kokomen/utils/general/analytics";
