+++ apps/kokomen-webview/src/domains/reports/components/feedbackAccordion.tsx
@@ -1,10 +1,5 @@
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger
-} from "@kokomen/ui/components/accordion";
-import { Score } from "@kokomen/ui/components/rank";
+import { Accordion } from "@kokomen/ui";
+import { Score } from "@kokomen/ui";
 import { MessageSquare, Star, Award } from "lucide-react";
 import { JSX } from "react";
 import { PrivateFeedback } from "@kokomen/types";
@@ -17,7 +12,7 @@ export function FeedbackAccordion({
   feedbacks: CamelCasedProperties<PrivateFeedback>[];
 }): JSX.Element {
   return (
-    <Accordion
+    <Accordion.Accordion
       allowMultiple
       defaultActiveKey={["feedback-1"]}
       className="w-full space-y-4"
@@ -29,7 +24,7 @@ export function FeedbackAccordion({
           idx={idx}
         />
       ))}
-    </Accordion>
+    </Accordion.Accordion>
   );
 }
 
@@ -41,12 +36,12 @@ function FeedBackAccordionItem({
   idx: number;
 }): JSX.Element {
   return (
-    <AccordionItem
+    <Accordion.AccordionItem
       key={feedback.questionId}
       itemKey={`feedback-${feedback.questionId}`}
       className="rounded-xl overflow-hidden bg-bg-elevated shadow-sm hover:shadow-md transition-shadow duration-200"
     >
-      <AccordionTrigger className="px-6 py-4 hover:bg-fill-secondary transition-colors duration-200">
+      <Accordion.AccordionTrigger className="px-6 py-4 hover:bg-fill-secondary transition-colors duration-200">
         <div className="flex items-center gap-3 w-full">
           <div className="flex items-center justify-center w-8 h-8 bg-primary-bg rounded-full">
             <span className="text-sm font-semibold text-primary">
@@ -59,8 +54,8 @@ function FeedBackAccordionItem({
             </p>
           </div>
         </div>
-      </AccordionTrigger>
-      <AccordionContent className="px-6">
+      </Accordion.AccordionTrigger>
+      <Accordion.AccordionContent className="px-6">
         <div className="flex flex-col gap-6">
           {/* 내 답변 섹션 */}
           <div className="space-y-3">
@@ -111,7 +106,7 @@ function FeedBackAccordionItem({
             visibility={feedback.answerMemoVisibility}
           />
         </div>
-      </AccordionContent>
-    </AccordionItem>
+      </Accordion.AccordionContent>
+    </Accordion.AccordionItem>
   );
 }
