+++ packages/ui/src/components/domains/index.ts
@@ -1,0 +1,1 @@
+export * from "./interviews";
