+++ packages/ui/src/components/domains/interviews/index.ts
@@ -1,0 +1,6 @@
+export { default as AiInterviewInterface } from "./AiInterviewInterface";
+export { default as InterviewFinishModal } from "./avatarMesh";
+export { default as CreateInterviewForm } from "./interviewBackground";
+export { default as InterviewInput } from "./interviewModals";
+export { default as InterviewSideBar } from "./interviewSideBar";
+export { default as Interviewer } from "./interviewer";
